requirements.txt
openai>=1.0.0
sounddevice
soundfile
pyserial
